==> heatwave/timeseries/timeseries_analysis/Scripts/1_assemble_physiological_data.Rmd <==
---
title: "Assemble master data frame for time series analysis"
author: "Ariana S Huffmyer, E5 RoL Team"
edited by: "DM Becker-Polinski"
date: "20220830"
output: github_document
editor_options: 
  chunk_output_type: console
--- 
 
 
# Set Up    
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message=FALSE)
```

```{r}
## install packages if you dont already have them
if (!require("tidyverse")) install.packages("tidyverse")
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("RColorBrewer")) install.packages("RColorBrewer")
if (!require("lme4")) install.packages("lme4")
if (!require("lmerTest")) install.packages("lmerTest")
if (!require("car")) install.packages("car")
if (!require("effects")) install.packages("effects")
if (!require("ggfortify")) install.packages("ggfortify")
if (!require("cowplot")) install.packages("cowplot")
if (!require("vegan")) install.packages("vegan")
if (!require("corrr")) install.packages("corrr")
if (!require("ggcorrplot")) install.packages("ggcorrplot")
if (!require("GGally")) install.packages("GGally")
if (!require("broom")) install.packages("broom")
if (!require("cowplot")) install.packages("cowplot")

# load packages
library(tidyverse)
library(ggplot2)
library(RColorBrewer)
library(lme4)
library(lmerTest)
library(car)
library(effects)
library(ggfortify)
library(cowplot)
library(vegan)
library(corrr)
library(ggcorrplot)
library(GGally)
library(broom)
library(cowplot)
```
 
# Load and manipulate data  

## Loading data files     

Load all .csv files from output of all timepoints for each physiological responses     

 
```{r}
biomass_files<-list.files("../", pattern = "biomass.calc.csv", recursive=T, full.names=T)
surface_area_files<-list.files("../", pattern = "surface.area.calc.csv", recursive=T, full.names=T)
host_protein_files<-list.files("../", pattern = "host_protein.csv", recursive=T, full.names=T)
holobiont_protein_files<-list.files("../", pattern = "holobiont_protein.csv", recursive=T, full.names=T)
symb_densities_files<-list.files("../", pattern = "sym_densities.csv", recursive=T, full.names=T)
chlorophyll_files<-list.files("../", pattern = "chlorophyll.csv", recursive=T, full.names=T)
```
 

## Read data files 
 
Load all data frames.  

```{r}
#biomass 

biomass_dataset <- data.frame()

for (i in 1:length(biomass_files)){
  biomass_df <- read.csv(biomass_files[i]) #each file will be read in
  biomass_dataset <- rbind(biomass_dataset, biomass_df) #for each iteration, bind the new data to the building dataset
}

```

```{r}
#surface area 
surface_area_dataset <- data.frame()

for (i in 1:length(surface_area_files)){
  surface_area_df <- read.csv(surface_area_files[i]) #each file will be read in
  surface_area_dataset <- rbind(surface_area_dataset, surface_area_df) #for each iteration, bind the new data to the building dataset
}
```

```{r}
#host protein 
host_protein_dataset <- data.frame()

for (i in 1:length(host_protein_files)){
  host_protein_df <- read.csv(host_protein_files[i]) #each file will be read in
  host_protein_dataset <- rbind(host_protein_dataset, host_protein_df) #for each iteration, bind the new data to the building dataset
}
```

```{r}
#holobiont protein 
holobiont_protein_dataset <- data.frame()

for (i in 1:length(holobiont_protein_files)){
  holobiont_protein_df <- read.csv(holobiont_protein_files[i]) #each file will be read in
  holobiont_protein_dataset <- rbind(holobiont_protein_dataset, holobiont_protein_df) #for each iteration, bind the new data to the building dataset
}
```

```{r}
#symb densities 
symb_densities_dataset <- data.frame()

for (i in 1:length(symb_densities_files)){
  symb_densities_df <- read.csv(symb_densities_files[i]) #each file will be read in
  symb_densities_dataset <- rbind(symb_densities_dataset, symb_densities_df) #for each iteration, bind the new data to the building dataset
}
```

```{r}
#chlorophyll files 
chlorophyll_dataset <- data.frame()

for (i in 1:length(chlorophyll_files)){
  chlorophyll_df <- read.csv(chlorophyll_files[i]) #each file will be read in
  chlorophyll_dataset <- rbind(chlorophyll_dataset, chlorophyll_df) #for each iteration, bind the new data to the building dataset
}

```



```{r}
#remove files that are not needed from loops
rm(list = ls(pattern = "*_df"))
```


## Generate master data frame    

Read in tag metadata frame. 
```{r}
#Load tag metadata sheet 
tags<-read.csv("fragment_ metadata.csv")
```

Prepare datasets for merging by renaming columns and spreading dataframes. Each file needs to have one line per colony per response variable in order to merge and should be without site/species columns as these will be added from metadata sheet.   
```{r}
# biomass data 
biomass_dataset<- biomass_dataset %>%
  nest(value_col = c(AFDW.mg.cm2, DW.mg.cm2)) %>%
  spread(key = partner, value = value_col) %>%
  unnest(Host, Sym, .sep = '_')%>%
  select(fragment_ID, Host_AFDW.mg.cm2, Sym_AFDW.mg.cm2, Host_DW.mg.cm2, Sym_DW.mg.cm2, timepoint)

#host protein data
host_protein_dataset<-host_protein_dataset%>%
  select(fragment_ID, prot_ug.cm2, timepoint)

#change host prot_ug.cm2 name to be spcific to host
host_protein_dataset <- host_protein_dataset %>%
  rename(host_prot_ug.cm2 = prot_ug.cm2)

#holobiont protein data
holobiont_protein_dataset<-holobiont_protein_dataset%>%
  select(fragment_ID, prot_ug.cm2, timepoint)

#change holobiont prot_ug.cm2 name to be spcific to holobiont
holobiont_protein_dataset <- holobiont_protein_dataset %>%
  rename(holobiont_prot_ug.cm2 = prot_ug.cm2)
  
#symbiont densitites
symb_densities_dataset<-symb_densities_dataset%>%
  select(fragment_ID, cells.cm2, timepoint)

```

Join all data frames together.  

```{r}
#can also use "join_all"
master <- full_join(full_join(full_join(full_join(full_join(
  chlorophyll_dataset,
  host_protein_dataset, by=c("fragment_ID", "timepoint")),
  holobiont_protein_dataset, by=c("fragment_ID", "timepoint")),
  surface_area_dataset, by=c("fragment_ID", "timepoint")), 
  biomass_dataset, by=c("fragment_ID", "timepoint")),
  symb_densities_dataset, by=c("fragment_ID", "timepoint"))

head(master)
```

Check for duplicates, will return TRUE if there are no duplicates.    
```{r}
master<-master%>%
  mutate(code=as.factor(paste(fragment_ID, "-", timepoint)))

length(unique(master$code)) == nrow(master)
length(unique(master$code))
nrow(master)

master<-master%>%
  unique()

length(unique(master$code)) == nrow(master)
length(unique(master$code))
nrow(master)

```

Export list of duplicate values to examine.  

```{r}
duplicates<-master[duplicated(master$code),]
duplicates<-duplicates%>%
   droplevels()
list<-levels(duplicates$code)
 
 master%>%
   filter(code %in% list)%>%
   write_csv(file="../output/timeseries_duplicates_qc.csv")
 

```
  

# Normalization and write master file  

In order to examine the effect of normalizer on all response variables, normalize each response to: 
(1) Host protein  (ug/cm2)  
(2) Host AFDW  (mg/cm2)   
(3) Host surface area (cm2)  

To do this, back-normalize variables that are already normalized and divide by each new normalizer.     

```{r}
master.1<-master%>%
  mutate(holobiont_mgprot = (holobiont_prot_ug.cm2*surface.area.cm2)/1000)%>% #get absolute values of normaliers for each coral sample, change from ug to mg
   mutate(host_mgprot = (host_prot_ug.cm2*surface.area.cm2)/1000)%>% #get absolute values of normaliers for each coral sample, change from ug to mg
 mutate(mgafdw = Host_AFDW.mg.cm2*surface.area.cm2)%>% #get absolute values of normaliers for each coral sample
  mutate(mgafdw = Sym_AFDW.mg.cm2*surface.area.cm2)%>% #get absolute values of normaliers for each coral sample
  mutate(cells = cells.cm2 * surface.area.cm2)%>% #get absolute number of symbiont cells
  mutate(cells.holobiont.mgprot = (cells.cm2 * surface.area.cm2)/holobiont_mgprot)%>% #convert cells per cm2 to cells per mg holobiont protein
  mutate(chla.ug.holobiont.mgprot = (chla.ug.cm2 * surface.area.cm2)/holobiont_mgprot)%>% #normalize chla to holobiont protein
  mutate(chlc2.ug.holobiont.mgprot = (chlc2.ug.cm2 * surface.area.cm2)/holobiont_mgprot)%>% #normalize chlc2 to holobiont protein
  mutate(holobiont_prot_mg.cm2=holobiont_prot_ug.cm2/1000)%>% #change holobiont protein to mg
  mutate(host_prot_mg.cm2=host_prot_ug.cm2/1000)%>% #change host protein to mg
  mutate(chla.ug.cell = (chla.ug.cm2 * surface.area.cm2)/cells) %>% #normalize to chl per cell
  mutate(chlc2.ug.cell = (chlc2.ug.cm2 * surface.area.cm2)/cells)  #normalize to chl per cell

```

# Calculate additional metrics of total cholorophyll (a + c2).  

```{r}
master.1 <- master.1 %>% 
 #mutate(Ratio_AFDW.mg.cm2=Sym_AFDW.mg.cm2/(Sym_AFDW.mg.cm2+Host_AFDW.mg.cm2))%>%
 #mutate(Total_Chl=chla.ug.mgAFDW+chlc2.ug.mgAFDW)%>%
 mutate(Total_Chl_cm2=chla.ug.cm2+chlc2.ug.cm2)

head(master.1)
```

 
Write master file to csv.  
```{r}

#merge metadata with master timeseries
metadata_master <- full_join(master.1, tags)

write_csv(metadata_master, "timeseries_analysis/Output/master_timeseries.csv")
```


# Generating summary by groups   

Summarize all response values by group (species + timepoint + site).       

```{r}

summary_fragments<-metadata_master%>%
  group_by(fragment_ID, treatment, timepoint)%>%
  summarise(across(.cols=everything(), ~mean(.x, na.rm = TRUE)))%>%
  drop_na("timepoint")%>%
  write_csv(., "timeseries_analysis/Output/summary_physio_fragment.csv")

summary_group<-metadata_master%>%
  group_by(treatment, timepoint)%>%
  summarise(across(.cols=everything(), ~mean(.x, na.rm = TRUE)))%>%
  drop_na("timepoint")%>%
  write_csv(., "timeseries_analysis/Output/summary_physio_timepoint.csv")
           
```
